Rename pruning methods and split modules
